refactor: moved Impiccato.py into main.py
